[수정] 포스트 리스트 페이징

+++ ohseatback/src/main/java/com/ohseat/ohseatback/mapper/RecommendMapper.java
@@ -24,7 +24,7 @@ public interface RecommendMapper {
             @Param("cinemaId") String cinemaId,
             @Param("screenId") String screenId,
             @Param("limit") int limit,
-            @Param("offset") int offset
+            @Param("offset") long offset
     );
 
     long countPosts(
